test(workflow): consolidate regression tests into existing test files
Move the three regression tests next to the code they cover instead of
adding new files: wrappedSession -> llmagent_saveoutput_test.go (the
white-box llmagent test), FunctionNode Content -> function_node_test.go,
and the runner MessageAsOutput test -> run_node_test.go.
